requirements.txt
appnope==0.1.0
cling==2.3.1
cookies==2.2.1
coverage==4.0.3
coveralls==1.1
decorator==4.0.9
dj-database-url==0.4.0
dj-static==0.0.6
Django==1.9.2
django-allauth==0.24.1
django-appconf==1.0.1
django-bootstrap3==6.2.2
django-facebook==6.0.3
django-imagekit==3.3
django-social-share==0.3.0
docopt==0.6.2
factory-boy==2.6.1
fake-factory==0.5.3
funcsigs==0.4
oauthlib==1.0.3
path.py==8.2.1
pilkit==1.1.13
Pillow==3.3.1
python-openid==2.2.5
requests==2.9.1
requests-oauthlib==0.6.0
six==1.10.0
static3==0.6.1
Unidecode==0.4.19
wheel==0.24.0
numpy
opencv
coverage
